defaults.inc.php: Update comments for smtp wirh TLS
The comment for smtp_conn_options suggested using smtp_timeout > 0 when using TLS.
The comment for smtp_timeout suggested that there is a bug when using 0 when using TLS. The references bug doesn't talk about smtp_fimeout at all, but that a server requiring SNI needs the connection parameter SNI_enabled.
I've uses TLS for smtp connections for several years and did use the default (=0) for smtp_timeout without problems. So let's remove the confusing and inconsistent comments and add SNI_enabled to the example.